==> 12_prediction_on_new_validation_data.ipynb <==
# continued from the models trained in 11_prepare_training_data_and_train.ipynb

import pandas as pd
import numpy as np
import pathlib 
import os
import matplotlib.pyplot as plt
import seaborn as sns
import random
from tqdm import tqdm
import warnings
import pandas as pd
import argparse
import pickle
from sklearn.linear_model import LinearRegression
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_squared_error, r2_score
from sklearn.linear_model import Ridge, Lasso, ElasticNet
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import roc_curve, auc

from dataset_paths import *

warnings.filterwarnings('ignore')


output_version = "20241229"
# outdir = "/media/hieunguyen/HNSD_mini/outdir"
outdir = "/media/hieunguyen/HNHD01/outdir"
dataset_name = "SPIKE_IN_20250122"
mode = "all"
PROJECT = "combine_ctcandi_ichorcna"

validation_dataset_name = "Truong_new_reqest_20250321"

for control_sample_mode in ["full", 25, 50, 75, "background_full" , "background_25", "background_50", "background_75", "background"]:
    for input_cancer_class in ["Liver", "Lung", "Breast", "CRC", "pan_cancer"]:
        for selected_candi_features in ["ratio_raw", "ratio_in_read"]:

            # control_sample_mode = "background_full"
            # input_cancer_class = "Liver"
            # selected_candi_features = "ratio_raw"

            input_features = ["ichorCNA", selected_candi_features] 

            path_to_main_output = os.path.join(outdir, PROJECT, output_version, dataset_name)
            path_to_07_output = os.path.join(path_to_main_output, f"07_output_{mode}", input_cancer_class)
            path_to_08_output = os.path.join(path_to_main_output, f"08_output_{mode}", input_cancer_class)
            path_to_11_output = os.path.join(path_to_main_output, 
                                                f"11_output_{mode}", 
                                                input_cancer_class, 
                                                f"feature_{selected_candi_features}", 
                                                f"{control_sample_mode}")
            path_to_12_output = os.path.join(path_to_main_output, 
                                                f"12_output_{mode}", 
                                                input_cancer_class, 
                                                f"feature_{selected_candi_features}", 
                                                f"{control_sample_mode}",
                                                validation_dataset_name)

            path_to_save_models = os.path.join(path_to_11_output, "model_files")

            # create the folder to save output in this analysis
            os.system(f"mkdir -p {path_to_12_output}")

            all_models = dict()
            get_all_models = [item for item in pathlib.Path(path_to_save_models).glob("*.pkl")]
            for model_file in get_all_models:
                model_name = model_file.name.replace(".pkl", "")
                tmp_model = pickle.load(open(str(model_file), "rb"))
                all_models[model_name] = tmp_model
            fulldf = pd.read_csv(os.path.join(outdir, PROJECT, output_version, validation_dataset_name,
                                            f"08_output_{mode}", input_cancer_class,
                                            "feature.csv"), index_col = [0])
            for modelid in all_models.keys():
                model = all_models[modelid]    
                fulldf[f"{modelid}_pred"] = model.predict(fulldf[input_features].to_numpy())

            fulldf.to_csv(os.path.join(path_to_12_output, "new_validation.csv"))
